refactor: 컴포넌트 분류 및 코드 개선
