+++ lib/home/component/background.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class Background extends StatelessWidget {
+  final Widget widgets;
+
+  const Background({required this.widgets, super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        body: Container(
+      width: MediaQuery.of(context).size.width,
+      height: MediaQuery.of(context).size.height,
+      decoration: BoxDecoration(
+        image: DecorationImage(
+          image: AssetImage("asset/background.png"),
+          fit: BoxFit.fill,
+        ),
+      ),
+      child: widgets,
+    ));
+  }
+}
