+++ lib/main.dart
@@ -1,6 +1,6 @@
-import 'package:ai_me/screen/home_screen.dart';
-import 'package:ai_me/screen/message_screen.dart';
-import 'package:ai_me/screen/result_screen.dart';
+import 'package:ai_me/home/home_screen.dart';
+import 'package:ai_me/message/message_screen.dart';
+import 'package:ai_me/result/result_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
@@ -9,7 +9,7 @@ void main() {
       ScreenUtilInit(designSize: Size(360, 690),
   builder:(context,child) {
     return MaterialApp(
-      home: MessageScreen(),
+      home: ResultScreen(),
     );
   }));
   }
